modular/requirements.txt
# Safety Copilot - Requirements
# Install with: pip install -r requirements.txt

# Core ML
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web UI
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0

# Optional: GPU support (uncomment for CUDA)
# --extra-index-url https://download.pytorch.org/whl/cu118
# torch>=2.0.0+cu118
# torchvision>=0.15.0+cu118
